Add Mx, a gender-neutral title
"But this array of options is still inadequate, because not everyone falls neatly into the binary model of gender. In official contexts we tend to categorise people as male/female, married/unmarried, ignoring the often more complex realities of identity. And just as Ms enables women not to indicate their marital status, an emerging title allows people not to indicate their gender: Mx."

http://www.macmillandictionaryblog.com/mx-a-new-gender-neutral-title