Burroughs computer opcodes
There is more of this to do... if this submission is approved I might gather the rest up too.

The Burroughs B6x00-7x00 instruction set includes the set of valid operations for the Burroughs computers manufactured in the 1970s. An instruction set is an abstract model of computer functions. In that era, distinct items in the set were called syllables or opcodes.

https://en.wikipedia.org/wiki/Burroughs_B6x00-7x00_instruction_set